+++ core/src/main/kotlin/com/falsepattern/zigbrains/project/settings/ZigProjectSettings.kt
@@ -26,6 +26,7 @@ import com.falsepattern.zigbrains.project.toolchain.LocalZigToolchain
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.util.io.toNioPathOrNull
 import com.intellij.util.xmlb.annotations.Transient
+import kotlin.io.path.isDirectory
 import kotlin.io.path.pathString
 
 data class ZigProjectSettings(
@@ -41,7 +42,13 @@ data class ZigProjectSettings(
     @get:Transient
     @set:Transient
     override var toolchain: LocalZigToolchain?
-        get() = toolchainPath?.toNioPathOrNull()?.let { LocalZigToolchain(it) }
+        get() {
+            val nioPath = toolchainPath?.toNioPathOrNull() ?: return null
+            if (!nioPath.isDirectory()) {
+                return null
+            }
+            return LocalZigToolchain(nioPath)
+        }
         set(value) {
             toolchainPath = value?.location?.pathString
         }