+++ core/src/main/kotlin/com/falsepattern/zigbrains/project/steps/discovery/ZigStepDiscoveryListener.kt
@@ -30,6 +30,7 @@ interface ZigStepDiscoveryListener {
 
     enum class ErrorType {
         MissingToolchain,
+        MissingZigExe,
         MissingBuildZig,
         GeneralError
     }