+++ core/src/main/kotlin/com/falsepattern/zigbrains/project/steps/ui/BuildToolWindowContext.kt
@@ -166,6 +166,7 @@ class BuildToolWindowContext(private val project: Project): Disposable {
             withEDTContext {
                 getViewport(project)?.setViewportError(ZigBrainsBundle.message(when(type) {
                     ZigStepDiscoveryListener.ErrorType.MissingToolchain -> "build.tool.window.status.error.missing-toolchain"
+                    ZigStepDiscoveryListener.ErrorType.MissingZigExe -> "build.tool.window.status.error.missing-zig-exe"
                     ZigStepDiscoveryListener.ErrorType.MissingBuildZig -> "build.tool.window.status.error.missing-build-zig"
                     ZigStepDiscoveryListener.ErrorType.GeneralError -> "build.tool.window.status.error.general"
                 }), details)