+++ core/src/main/kotlin/com/falsepattern/zigbrains/project/toolchain/tools/ZigTool.kt
@@ -31,12 +31,13 @@ import kotlinx.coroutines.runInterruptible
 import kotlinx.coroutines.withContext
 import kotlinx.coroutines.withTimeoutOrNull
 import java.nio.file.Path
+import kotlin.io.path.isRegularFile
 
 abstract class ZigTool(val toolchain: AbstractZigToolchain) {
     abstract val toolName: String
 
-    suspend fun callWithArgs(workingDirectory: Path?, vararg parameters: String, timeoutMillis: Long = Long.MAX_VALUE): ProcessOutput {
-        val cli = createBaseCommandLine(workingDirectory, *parameters)
+    suspend fun callWithArgs(workingDirectory: Path?, vararg parameters: String, timeoutMillis: Long = Long.MAX_VALUE): ProcessOutput? {
+        val cli = createBaseCommandLine(workingDirectory, *parameters) ?: return null
 
         val (process, exitCode) = withContext(Dispatchers.IO) {
             val process = cli.createProcess()
@@ -59,9 +60,12 @@ abstract class ZigTool(val toolchain: AbstractZigToolchain) {
     private suspend fun createBaseCommandLine(
         workingDirectory: Path?,
         vararg parameters: String
-    ): GeneralCommandLine {
+    ): GeneralCommandLine? {
+        val exe = toolchain.pathToExecutable(toolName)
+        if (!exe.isRegularFile())
+            return null
         val cli = GeneralCommandLine()
-            .withExePath(toolchain.pathToExecutable(toolName).toString())
+            .withExePath(exe.toString())
             .withWorkingDirectory(workingDirectory)
             .withParameters(*parameters)
             .withCharset(Charsets.UTF_8)